add regular siding, horizontal and vertical, add slope cutting and pitch offset
